THREESCALE-10842 address comments p3
