+++ pkg/helper/requirements_common_helper_test.go
@@ -11,8 +11,8 @@ func TestVersionCompare(t *testing.T) {
 		incomingVersion string
 		expectedResult  bool
 	}{
-		{"IncomingMajorRequiredIsHiger", "7.0.0", "6.2", false},
-		{"IncomingMinorRequiredIsHiger", "6.3.0", "6.2", false},
+		{"IncomingMajorRequiredIsHigher", "7.0.0", "6.2", false},
+		{"IncomingMinorRequiredIsHigher", "6.3.0", "6.2", false},
 		{"IncomingMajorRequiredIsLower", "5.0.0", "6.0.0", true},
 		{"IncomingMinorRequiredIsLower", "6.1.0", "6.2", true},
 		{"VersionsMatch", "6.2.0", "6.2.0", true},
